backend/requirements.txt
asgiref==3.6.0
Django==4.1.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.10
pytz==2023.3
setuptools==80.9.0
sqlparse==0.4.4
